+++ server.js
@@ -5,6 +5,11 @@ const postsRouter = require( './posts/postRouter' );
 
 const server = express();
 
+function logger( req, res, next ) {
+  console.log( `${ req.method }to ${ req.originalUrl }` )
+  next();
+}
+
 server.get('/', (req, res) => {
   res.send(`<h2>Let's write some middleware!</h2>`);
 });
@@ -13,6 +18,7 @@ server.get('/', (req, res) => {
 
 server.use( '/users', userRouter );
 server.use( '/posts', postsRouter );
-
+server.use( logger );
+server.use( express.json() );
 
 module.exports = server;
